feat: 예외 처리

+++ backend/src/main/java/com/yukgaejang/voss/domain/freeboard/service/PostCommentServiceImpl.java
@@ -1,8 +1,10 @@
 package com.yukgaejang.voss.domain.freeboard.service;
 
 import com.yukgaejang.voss.domain.freeboard.exception.NoPostCommentException;
+import com.yukgaejang.voss.domain.freeboard.exception.NoPostException;
 import com.yukgaejang.voss.domain.freeboard.repository.PostCommentRepository;
 import com.yukgaejang.voss.domain.freeboard.repository.PostRepository;
+import com.yukgaejang.voss.domain.freeboard.repository.entity.Post;
 import com.yukgaejang.voss.domain.freeboard.repository.entity.PostComment;
 import com.yukgaejang.voss.domain.freeboard.service.dto.request.CreateCommentRequest;
 import com.yukgaejang.voss.domain.freeboard.service.dto.request.UpdateCommentRequest;
@@ -10,6 +12,7 @@ import com.yukgaejang.voss.domain.freeboard.service.dto.response.CommentDetailRe
 import com.yukgaejang.voss.domain.freeboard.service.dto.response.CreateCommentResponse;
 import com.yukgaejang.voss.domain.freeboard.service.dto.response.DeleteCommentResponse;
 import com.yukgaejang.voss.domain.freeboard.service.dto.response.UpdateCommentResponse;
+import com.yukgaejang.voss.domain.member.exception.NoMemberException;
 import com.yukgaejang.voss.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -30,9 +33,11 @@ public class PostCommentServiceImpl implements PostCommentService {
 
     @Override
     public CreateCommentResponse createComment(Long postId, String email, CreateCommentRequest createCommentRequest) {
+        Post post = postRepository.findByIdAndIsDeletedFalse(postId);
+        if(post == null) throw new NoPostException("존재하지 않는 글입니다.");
         PostComment postComment = PostComment.builder()
-                .member(memberRepository.findByEmail(email).orElseThrow(() -> new RuntimeException("존재하지 않는 사용자입니다.")))
-                .post(postRepository.findByIdAndIsDeletedFalse(postId))
+                .member(memberRepository.findByEmail(email).orElseThrow(() -> new NoMemberException("존재하지 않는 사용자입니다.")))
+                .post(post)
                 .content(createCommentRequest.getContent())
                 .build();
         postCommentRepository.save(postComment);
@@ -40,7 +45,9 @@ public class PostCommentServiceImpl implements PostCommentService {
     }
 
     @Override
-    public UpdateCommentResponse updateComment(Long commentId, UpdateCommentRequest updateCommentRequest) {
+    public UpdateCommentResponse updateComment(Long postId, Long commentId, UpdateCommentRequest updateCommentRequest) {
+        Post post = postRepository.findByIdAndIsDeletedFalse(postId);
+        if(post == null) throw new NoPostException("존재하지 않는 글입니다.");
         PostComment postComment = postCommentRepository.findByIdAndIsDeletedFalse(commentId);
         if(postComment == null) throw new NoPostCommentException("존재하지 않는 댓글입니다.");
         postComment.update(updateCommentRequest.getContent());
@@ -50,11 +57,15 @@ public class PostCommentServiceImpl implements PostCommentService {
 
     @Override
     public List<CommentDetailResponse> getComments(Long postId) {
+        Post post = postRepository.findByIdAndIsDeletedFalse(postId);
+        if(post == null) throw new NoPostException("존재하지 않는 글입니다.");
         return postCommentRepository.findAllByPostIdAndIsDeletedFalse(postId);
     }
 
     @Override
-    public DeleteCommentResponse deleteComment(Long commentId) {
+    public DeleteCommentResponse deleteComment(Long postId, Long commentId) {
+        Post post = postRepository.findByIdAndIsDeletedFalse(postId);
+        if(post == null) throw new NoPostException("존재하지 않는 글입니다.");
         PostComment postComment = postCommentRepository.findByIdAndIsDeletedFalse(commentId);
         if(postComment == null) throw new NoPostCommentException("존재하지 않는 댓글입니다.");
         postComment.delete();
